Refactor ProductsController and ProductsService
This commit refactors the ProductsController and ProductsService classes.

In the ProductsController class, a new method called validateProduct has been added. This method is decorated with @MessagePattern and is used to validate products based on their IDs. The method takes an array of IDs as input and returns the corresponding products after validation. If any of the products are not found, a RpcException with a status code of HttpStatus.BAD_REQUEST and a message indicating that some products were not found is thrown.

In the ProductsService class, the OnModuleInit interface has been removed from the class declaration. Additionally, a new method called validateProducts has been added. This method takes an array of product IDs as input, removes any duplicate IDs, and queries the database to find the corresponding products. If any of the products are not found, a RpcException with a status code of HttpStatus.BAD_REQUEST and a message indicating that some products were not found is thrown.

These changes improve the functionality and maintainability of the ProductsController and ProductsService classes.